Move auth to library
