+++ apps/api-gateway/src/app.module.ts
@@ -2,9 +2,8 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule } from '@nestjs/config';
-import { AuthModule } from './features/auth/auth.module';
+import { AuthModule, AuthGuard } from '@app/auth';
 import { APP_GUARD } from '@nestjs/core';
-import { AuthGuard } from './features/auth/guards/auth.guard';
 import { KafkaModule } from '@app/kafka';
 import { UserModule } from './features/user/user.module';
 import { TargetModule } from './features/target/target.module';
