+++ libs/auth/src/auth.module.ts
@@ -5,7 +5,7 @@ import { JwtStrategy } from './strategies/jwt.strategy';
 import { ConfigModule } from '@nestjs/config';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
-import { UserModule } from '../user/user.module';
+import { UserModule } from '../../../apps/api-gateway/src/features/user/user.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 @Module({
