+++ libs/auth/src/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './user.dto';
