+++ libs/auth/src/dto/user.dto.ts
