+++ libs/auth/src/guards/auth.guard.ts
@@ -7,8 +7,8 @@ import {
 import { JwtService } from '@nestjs/jwt';
 import { Request } from 'express';
 import { Reflector } from '@nestjs/core';
-import { IS_PUBLIC_KEY } from '../../../common/decorators/public';
-import { UserService } from '../../../features/user/user.service';
+import { IS_PUBLIC_KEY } from '../../../../apps/api-gateway/src/common/decorators/public';
+import { UserService } from '../../../../apps/api-gateway/src/features/user/user.service';
 
 @Injectable()
 export class AuthGuard implements CanActivate {
