+++ libs/auth/src/index.ts
@@ -1,0 +1,7 @@
+export * from './auth.controller';
+export * from './auth.module';
+export * from './auth.service';
+export * from './dto/index';
+export * from './guards/index';
+export * from './strategies/index';
+export * from './utils/index';
