+++ libs/auth/src/strategies/index.ts
@@ -1,0 +1,1 @@
+export * from './jwt.strategy';
