+++ libs/auth/src/strategies/jwt.strategy.ts
