+++ package.json
@@ -92,6 +92,7 @@
       "<rootDir>/libs/"
     ],
     "moduleNameMapper": {
+      "^@app/auth(|/.*)$": "<rootDir>/libs/auth/src/$1",
       "^@app/kafka(|/.*)$": "<rootDir>/libs/kafka/src/$1",
       "^@app/types(|/.*)$": "<rootDir>/libs/types/src/$1"
     }
